feat: organize CLI files
